IntNodes: moved from tilelink to their own package
